Adapt the manuscript examples (quine, etc) to 32-bit.

+++ manuscript/code/countdown.c
@@ -1,9 +1,9 @@
-// declaring bar as a global variable of type uint64_t and
+// declaring bar as a global variable of type uint32_t and
 // defining the decimal value 10 as initial value of bar
-uint64_t bar = 10;
+uint32_t bar = 10;
 
-// declaring main as a procedure with no arguments and return type uint64_t
-uint64_t main() {
+// declaring main as a procedure with no arguments and return type uint32_t
+uint32_t main() {
   // defining main to be the following code
 
   // while bar is greater than the decimal value 0 ...
