+++ manuscript/code/iteration.c
@@ -1,6 +1,6 @@
-uint64_t x;
+uint32_t x;
 
-uint64_t main() {
+uint32_t main() {
   x = 0;
 
   x = x + 1;
