+++ manuscript/code/local.c
@@ -1,12 +1,12 @@
-uint64_t f(uint64_t x) {
+uint32_t f(uint32_t x) {
   while (x > 0)
     x = x - 1;
 
   return x;
 }
 
-uint64_t main() {
-  uint64_t x;
+uint32_t main() {
+  uint32_t x;
   
   x = 0;
